move decl tests to decl.rs
