+++ ecmascript/codegen/src/decl.rs
@@ -102,3 +102,27 @@ impl<'a> Emitter<'a> {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::tests::assert_min;
+
+    #[test]
+    fn issue_275() {
+        assert_min(
+            "function* foo(){
+            yield getServiceHosts()
+        }",
+            "function*foo(){yield getServiceHosts();}",
+        );
+    }
+
+    #[test]
+    fn single_argument_arrow_expression() {
+        assert_min("function* f(){ yield x => x}", "function*f(){yield x=>x;}");
+        assert_min(
+            "function* f(){ yield ({x}) => x}",
+            "function*f(){yield({x})=>x;}",
+        );
+    }
+}
